Break up texturenode.c into individual texture node files.
More code organization churn, but we're getting there!
