+++ src/nodes/shaders/bsdf.h
@@ -11,18 +11,7 @@
 #include "../../datatypes/lightRay.h"
 #include "../../utils/mempool.h"
 #include "../textures/texturenode.h"
-
-//FIXME: Move
-struct hitRecord {
-	struct lightRay incident;		//Light ray that encountered this intersection
-	struct material material;		//Material of the intersected object
-	struct vector hitPoint;			//Hit point vector in 3D space
-	struct vector surfaceNormal;	//Surface normal at that point of intersection
-	struct coord uv;				//UV barycentric coordinates for intersection point
-	float distance;					//Distance to intersection point
-	struct poly *polygon;			//ptr to polygon that was encountered
-	int instIndex;					//Instance index, negative if no intersection
-};
+#include "../../datatypes/hitrecord.h"
 
 struct bsdfSample {
 	struct vector out;
