+++ src/nodes/textures/constant.h
@@ -1,0 +1,11 @@
+//
+//  constant.h
+//  C-Ray
+//
+//  Created by Valtteri Koskivuori on 06/12/2020.
+//  Copyright © 2020 Valtteri Koskivuori. All rights reserved.
+//
+
+#pragma once
+
+struct textureNode *newConstantTexture(struct block **pool, struct color color);
