+++ src/nodes/textures/image.h
@@ -1,0 +1,14 @@
+//
+//  image.h
+//  C-Ray
+//
+//  Created by Valtteri Koskivuori on 06/12/2020.
+//  Copyright © 2020 Valtteri Koskivuori. All rights reserved.
+//
+
+#pragma once
+
+#define SRGB_TRANSFORM 0x01
+#define NO_BILINEAR    0x02
+
+struct textureNode *newImageTexture(struct block **pool, struct texture *texture, uint8_t options);
