+++ src/nodes/textures/texturenode.h
@@ -8,8 +8,6 @@
 
 #pragma once
 
-//TODO: Break this up into a color node, then expanding on that.
-
 enum textureType {
 	Diffuse,
 	Normal,
@@ -20,13 +18,6 @@ struct textureNode {
 	struct color (*eval)(const struct textureNode *node, const struct hitRecord *record);
 };
 
-#define SRGB_TRANSFORM 0x01
-#define NO_BILINEAR    0x02
-
-struct textureNode *newImageTexture(struct block **pool, struct texture *texture, uint8_t options);
-
-struct textureNode *newConstantTexture(struct block **pool, struct color color);
-
-struct textureNode *newColorCheckerBoardTexture(struct block **pool, struct textureNode *colorA, struct textureNode *colorB, float size);
-
-struct textureNode *newCheckerBoardTexture(struct block **pool, float size);
+#include "checker.h"
+#include "constant.h"
+#include "image.h"
